requirements.txt
langchain
langchain-community
langchain-openai
langchain-huggingface
langchain-ollama
langchain-chroma
langchain-text-splitters

python-dotenv
pypdf
beautifulsoup4
chromadb
faiss-cpu
sentence-transformers
ipykernel
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
bs4
langchain_ollama
langgraph
streamlit
arxiv
wikipedia
huggingface_hub
tiktoken
datasets
cassio
gradio
ddgs